fixed include of ElectronStream plugin, moving part of include statements into .cc file and removed some of the unuseful includes
